+++ core/src/main/java/com/graphhopper/routing/util/parsers/BikeAccessParser.java
@@ -9,13 +9,14 @@ import com.graphhopper.util.PMap;
 public class BikeAccessParser extends BikeCommonAccessParser {
 
     public BikeAccessParser(EncodedValueLookup lookup, PMap properties) {
-        this(lookup.getBooleanEncodedValue(VehicleAccess.key(properties.getString("name", "bike"))));
+        this(lookup.getBooleanEncodedValue(VehicleAccess.key(properties.getString("name", "bike"))),
+                lookup.getBooleanEncodedValue(Roundabout.KEY));
         blockPrivate(properties.getBool("block_private", true));
         blockFords(properties.getBool("block_fords", false));
     }
 
-    public BikeAccessParser(BooleanEncodedValue accessEnc) {
-        super(accessEnc);
+    public BikeAccessParser(BooleanEncodedValue accessEnc, BooleanEncodedValue roundaboutEnc) {
+        super(accessEnc, roundaboutEnc);
         barriers.add("kissing_gate");
         barriers.add("stile");
         barriers.add("turnstile");