+++ core/src/main/java/com/graphhopper/routing/util/parsers/MountainBikeAccessParser.java
@@ -9,13 +9,14 @@ import com.graphhopper.util.PMap;
 public class MountainBikeAccessParser extends BikeCommonAccessParser {
 
     public MountainBikeAccessParser(EncodedValueLookup lookup, PMap properties) {
-        this(lookup.getBooleanEncodedValue(VehicleAccess.key(properties.getString("name", "mtb"))));
+        this(lookup.getBooleanEncodedValue(VehicleAccess.key(properties.getString("name", "mtb"))),
+                lookup.getBooleanEncodedValue(Roundabout.KEY));
         blockPrivate(properties.getBool("block_private", true));
         blockFords(properties.getBool("block_fords", false));
     }
 
-    protected MountainBikeAccessParser(BooleanEncodedValue accessEnc) {
-        super(accessEnc);
+    protected MountainBikeAccessParser(BooleanEncodedValue accessEnc, BooleanEncodedValue roundaboutEnc) {
+        super(accessEnc, roundaboutEnc);
     }
 
     @Override