+++ core/src/main/java/com/graphhopper/routing/util/parsers/RacingBikeAccessParser.java
@@ -9,13 +9,14 @@ import com.graphhopper.util.PMap;
 public class RacingBikeAccessParser extends BikeCommonAccessParser {
 
     public RacingBikeAccessParser(EncodedValueLookup lookup, PMap properties) {
-        this(lookup.getBooleanEncodedValue(VehicleAccess.key(properties.getString("name", "racingbike"))));
+        this(lookup.getBooleanEncodedValue(VehicleAccess.key(properties.getString("name", "racingbike"))),
+                lookup.getBooleanEncodedValue(Roundabout.KEY));
         blockPrivate(properties.getBool("block_private", true));
         blockFords(properties.getBool("block_fords", false));
     }
 
-    protected RacingBikeAccessParser(BooleanEncodedValue accessEnc) {
-        super(accessEnc);
+    protected RacingBikeAccessParser(BooleanEncodedValue accessEnc, BooleanEncodedValue roundaboutEnc) {
+        super(accessEnc, roundaboutEnc);
 
         barriers.add("kissing_gate");
         barriers.add("stile");